fix(residue): 修复 win32 ^5.5.4 下注册表/快捷方式扫描的 FFI 签名错配
- `RegQueryValueEx` 的 `lpData` 形参在 win32 5.5.4 为 `Pointer<Uint8>`，
  原先在 `readDword` 里直接传 `Pointer<DWORD>` 会触发静态类型错误，
  导致整个项目无法编译；改为 `data.cast<Uint8>()`。
- `SLGP_RAWPATH` 在该版本未导出常量；用本地常量 `0x0004` 替代，
  避免 `shortcut_resolver` 编译失败。

这两处是卸载残留功能可编译并能跑通的前置修复。修复后已在真实注册表下
验证 `InstalledProgramsRepository.build()` + `RegistryResidueScanner.scan()`
正常完成扫描。

+++ lib/data/platform/registry_reader.dart
@@ -234,7 +234,14 @@ class RegKey {
     final data = calloc<DWORD>();
     final size = calloc<DWORD>()..value = sizeOf<DWORD>();
     try {
-      final rc = RegQueryValueEx(_handle, lpName, nullptr, type, data, size);
+      final rc = RegQueryValueEx(
+        _handle,
+        lpName,
+        nullptr,
+        type,
+        data.cast<Uint8>(),
+        size,
+      );
       if (rc != ERROR_SUCCESS) return null;
       if (type.value != REG_DWORD) return null;
       return data.value;
